+++ src/main/java/nextstep/repository/ReservationJdbcTemplateDao.java
@@ -1,0 +1,86 @@
+package nextstep.repository;
+
+import nextstep.domain.Reservation;
+import nextstep.domain.Theme;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
+
+import java.sql.Date;
+import java.sql.PreparedStatement;
+import java.sql.Time;
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+@Repository
+public class ReservationJdbcTemplateDao implements ReservationDao {
+    private JdbcTemplate jdbcTemplate;
+
+    public ReservationJdbcTemplateDao() {
+    }
+
+    @Autowired
+    public ReservationJdbcTemplateDao(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public Long save(Reservation reservation) {
+        final String insertSql = "INSERT INTO reservation (date, time, name, theme_name, theme_desc, theme_price) VALUES (?, ?, ?, ?, ?, ?);";
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            PreparedStatement ps = connection.prepareStatement(insertSql, new String[]{"id"});
+            ps.setDate(1, Date.valueOf(reservation.getDate()));
+            ps.setTime(2, Time.valueOf(reservation.getTime()));
+            ps.setString(3, reservation.getName());
+            ps.setString(4, reservation.getTheme().getName());
+            ps.setString(5, reservation.getTheme().getDesc());
+            ps.setInt(6, reservation.getTheme().getPrice());
+            return ps;
+        }, keyHolder);
+
+        return keyHolder.getKey().longValue();
+    }
+
+    public int countByDateAndTime(LocalDate date, LocalTime time) {
+        String sql = "SELECT count(*) FROM reservation WHERE date = ? and time = ?";
+        return jdbcTemplate.queryForObject(sql, Integer.class, date, time);
+    }
+
+    public Reservation findById(Long id) {
+        String sql = "SELECT * FROM reservation WHERE id = ?";
+        try {
+            return jdbcTemplate.queryForObject(sql, (rs, rowNum) ->
+                    new Reservation(
+                            rs.getLong("id"),
+                            rs.getDate("date").toLocalDate(),
+                            rs.getTime("time").toLocalTime(),
+                            rs.getString("name"),
+                            new Theme(
+                                    rs.getString("theme_name"),
+                                    rs.getString("theme_desc"),
+                                    rs.getInt("theme_price"))
+                    ), id);
+        } catch (EmptyResultDataAccessException e) {
+            return null;
+        }
+    }
+
+    public void clear() {
+        String sql = "DELETE FROM reservation";
+        jdbcTemplate.update(sql);
+        resetId();
+    }
+
+    private void resetId() {
+        String sql = "ALTER TABLE reservation ALTER COLUMN id RESTART WITH 1";
+        jdbcTemplate.execute(sql);
+    }
+
+    public void delete(Long id) {
+        String sql = "DELETE FROM reservation WHERE id = ?";
+        jdbcTemplate.update(sql, id);
+    }
+}
